refactor: article 도메인과 관련된 컴포넌트 components 밑의 article 폴더 밑으로 모음
