+++ frontend/src/components/article/PopularArticle/PopularArticle.tsx
@@ -1,8 +1,8 @@
 import EmptyMessage from '@/components/@common/EmptyMessage/EmptyMessage';
 import Loading from '@/components/@common/Loading/Loading';
+import * as S from '@/components/article/PopularArticle/PopularArticle.styles';
+import PopularArticleItem from '@/components/article/PopularArticleItem/PopularArticleItem';
 import useGetPopularArticles from '@/hooks/article/useGetPopularArticles';
-import * as S from '@/pages/Home/PopularArticle/PopularArticle.styles';
-import PopularArticleItem from '@/pages/Home/PopularArticleItem/PopularArticleItem';
 import { convertIdxToArticleColorKey } from '@/utils/converter';
 
 const PopularArticle = () => {
