+++ frontend/src/pages/CategoryArticles/CategoryArticles.tsx
@@ -1,10 +1,10 @@
 import { useNavigate, useParams } from 'react-router-dom';
 
-import ArticleItem from '@/components/@common/ArticleItem/ArticleItem';
 import EmptyMessage from '@/components/@common/EmptyMessage/EmptyMessage';
 import Loading from '@/components/@common/Loading/Loading';
 import ResponsiveInfiniteCardList from '@/components/@common/ResponsiveInfiniteCardList/ResponsiveInfiniteCardList';
 import SortDropdown from '@/components/@common/SortDropdown/SortDropDown';
+import ArticleItem from '@/components/article/ArticleItem/ArticleItem';
 import { URL } from '@/constants/url';
 import useGetCategoryArticles from '@/hooks/article/useGetCategoryArticles';
 import * as S from '@/pages/CategoryArticles/CategoryArticles.styles';
